This commit strengthens the typing in SGS.
Especially for what's exported. First and foremost it adds a type to the
`setup` directive, which is where most of the work dapp developers will
be doing will be. This should help them build apps dramatically faster
and with more confidence.

The commit also exports the types that shardus uses internally (those
that shardus exports, of course). This'll help dapp developers not have
to reinvent the wheel and create their own types that are an
approximation of the types shardus uses internally.

The commit also moves some account data types that were originally `any`
to the more accurage `unknown`. If we have no idea what the type will
be, it's more accurate to call it unknown than any.

Some fields in `shardus.setup` are made optional - the ones that I've
never seen before or aren't included in the dapps I've built so far, so
I'm sure they are indeed optional since those dapps work.

Finally there's some whitespace cleanup in here. My editor does it
automatically and I don't want to add each line individually. Errant
whitespace should not exist in the codebase anyways.