Update build-index.d.ts in attempt to export types from server-dist
This is somewhat of a shot in the dark, I want typing to be exported
from SGS when used as a library by a dapp developer. This is the
thirdish commit attempting to do that, and we haven't seen success yet.
However, I believe this could be the one!

Furthermore, in exploring this, I've become more familiar with this
build process, which I call a victory :)