Add a gitlab pipeline to check typescript/dependencies
Add a pipeline: This is a new thing for this project, so bing
bang boom here's one.

Check typescript: This makes sure typescript compiles, lest a small red
X appear on the project page.

Check dependencies: This also makes sure there are no extraneous
dependencies listed in the project that aren't being used. To get
this in, we had to remove two big ones, so this is already doing
good things.

This addresses
https://www.notion.so/shardus/a0e84bf31b97437eac5747d7940fb1c3?v=3bb4c72041194bb08a9fd5b3fcadb659&p=0bfd4f5f46754f94ae0c30c1430fe501