Add a note about installation in README
It's always been the case that we need to have cargo installed to
install this repo. However, it's never been the case that that fact has
been documented. This commit documents it.

We have tentative plans for the future to distribute prebuilt rust
binaries which will allow us to install this without the rust toolchain.
However, that's in the future, and this is now.