feat: Add minNodesPerConsensusGroup configuration option
This commit adds a new configuration option `minNodesPerConsensusGroup` to the `ServerConfiguration` interface. This option allows specifying the minimum number of nodes that should be present in a consensus group at any point in time.

The `minNodesPerConsensusGroup` option is added to the `sharding` object in the `SERVER_CONFIG` constant in the `src/config/server.ts` file. It is also used in the `StateManager` class in the `src/state-manager/index.ts` file and the `TransactionQueue` class in the `src/state-manager/TransactionQueue.ts` file to calculate the number of nodes in a consensus group.

This change improves the reliability and safety of the consensus algorithm by ensuring that there are always a minimum number of nodes participating in the consensus process.